finish session24-25

+++ session24-restAPI-server/controllers/feed.js
@@ -131,6 +131,12 @@ exports.updatePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error("Not authorized");
+				error.statusCode = 403;
+				throw error;
+			}
 			if (!imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
@@ -160,9 +166,21 @@ exports.deletePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error("Not authorized");
+				error.statusCode = 403;
+				throw error;
+			}
 			//CHeck logged in user
 			clearImage(post.imageUrl);
 			return Post.findByIdAndRemove(postId);
+		})
+		.then(result => {
+			return User.findById(req.userId);
+		})
+		.then(user => {
+			user.posts.pull(postId);
+			return user.save();
 		})
 		.then(result => {
 			console.log(result);
